+++ khoros/objects/__init__.py
@@ -6,12 +6,13 @@
 :Example:           ``node_id = base.get_node_id('https://community.khoros.com/t5/Khoros-Blog/bg-p/relnote', 'blog')``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     07 May 2020
+:Modified Date:     11 May 2020
 """
 
-__all__ = ['albums', 'messages', 'users']
+__all__ = ['albums', 'messages', 'roles', 'users']
 
 # Import all submodules by default
 from . import albums
 from . import messages
+from . import roles
 from . import users